Reorganize clustering benchmarks
