+++ benchmarks/clustering/dbscan.cpp
@@ -9,8 +9,8 @@
  * SPDX-License-Identifier: BSD-3-Clause                                    *
  ****************************************************************************/
 
-#include "dbscan.hpp"
-
+#include "ArborX_DBSCANVerification.hpp"
+#include <ArborX_DBSCAN.hpp>
 #include <ArborX_Version.hpp>
 
 #include <Kokkos_Core.hpp>
@@ -23,34 +23,211 @@
 #include <sstream>
 #include <vector>
 
-// FIXME: ideally, this function would be next to `loadData` in
-// dbscan_timpl.hpp. However, that file is used for explicit instantiation,
-// which would result in multiple duplicate symbols. So it is kept here.
-int getDataDimension(std::string const &filename, bool binary)
+#include "data.hpp"
+#include "parameters.hpp"
+#include "print_timers.hpp"
+
+template <typename MemorySpace>
+void writeLabelsData(std::string const &filename,
+                     Kokkos::View<int *, MemorySpace> labels)
 {
-  std::ifstream input;
-  if (!binary)
-    input.open(filename);
-  else
-    input.open(filename, std::ifstream::binary);
-  if (!input.good())
-    throw std::runtime_error("Error reading file \"" + filename + "\"");
+  std::ofstream out(filename, std::ofstream::binary);
+  ARBORX_ASSERT(out.good());
 
-  int num_points;
-  int dim;
-  if (!binary)
-  {
-    input >> num_points;
-    input >> dim;
-  }
-  else
-  {
-    input.read(reinterpret_cast<char *>(&num_points), sizeof(int));
-    input.read(reinterpret_cast<char *>(&dim), sizeof(int));
-  }
-  input.close();
+  auto labels_host =
+      Kokkos::create_mirror_view_and_copy(Kokkos::HostSpace{}, labels);
 
-  return dim;
+  int n = labels_host.size();
+  out.write((char *)&n, sizeof(int));
+  out.write((char *)labels_host.data(), sizeof(int) * n);
+}
+
+template <typename ExecutionSpace, typename LabelsView,
+          typename ClusterIndicesView, typename ClusterOffsetView>
+void sortAndFilterClusters(ExecutionSpace const &exec_space,
+                           LabelsView const &labels,
+                           ClusterIndicesView &cluster_indices,
+                           ClusterOffsetView &cluster_offset,
+                           int cluster_min_size = 1)
+{
+  Kokkos::Profiling::pushRegion("ArborX::DBSCAN::sortAndFilterClusters");
+
+  namespace KokkosExt = ArborX::Details::KokkosExt;
+
+  static_assert(Kokkos::is_view<LabelsView>{});
+  static_assert(Kokkos::is_view<ClusterIndicesView>{});
+  static_assert(Kokkos::is_view<ClusterOffsetView>{});
+
+  using MemorySpace = typename LabelsView::memory_space;
+
+  static_assert(std::is_same<typename LabelsView::value_type, int>{});
+  static_assert(std::is_same<typename ClusterIndicesView::value_type, int>{});
+  static_assert(std::is_same<typename ClusterOffsetView::value_type, int>{});
+
+  static_assert(std::is_same<typename LabelsView::memory_space, MemorySpace>{});
+  static_assert(
+      std::is_same<typename ClusterIndicesView::memory_space, MemorySpace>{});
+  static_assert(
+      std::is_same<typename ClusterOffsetView::memory_space, MemorySpace>{});
+
+  ARBORX_ASSERT(cluster_min_size >= 1);
+
+  int const n = labels.extent_int(0);
+
+  Kokkos::View<int *, MemorySpace> cluster_sizes(
+      "ArborX::DBSCAN::cluster_sizes", n);
+  Kokkos::parallel_for(
+      "ArborX::DBSCAN::compute_cluster_sizes",
+      Kokkos::RangePolicy(exec_space, 0, n), KOKKOS_LAMBDA(int const i) {
+        // Ignore noise points
+        if (labels(i) < 0)
+          return;
+
+        Kokkos::atomic_inc(&cluster_sizes(labels(i)));
+      });
+
+  // This kernel serves dual purpose:
+  // - it constructs an offset array through exclusive prefix sum, with a
+  //   caveat that small clusters (of size < cluster_min_size) are filtered out
+  // - it creates a mapping from a cluster index into the cluster's position in
+  //   the offset array
+  // We reuse the cluster_sizes array for the second, creating a new alias for
+  // it for clarity.
+  auto &map_cluster_to_offset_position = cluster_sizes;
+  constexpr int IGNORED_CLUSTER = -1;
+  int num_clusters;
+  KokkosExt::reallocWithoutInitializing(exec_space, cluster_offset, n + 1);
+  Kokkos::parallel_scan(
+      "ArborX::DBSCAN::compute_cluster_offset_with_filter",
+      Kokkos::RangePolicy(exec_space, 0, n),
+      KOKKOS_LAMBDA(int const i, int &update, bool final_pass) {
+        bool is_cluster_too_small = (cluster_sizes(i) < cluster_min_size);
+        if (!is_cluster_too_small)
+        {
+          if (final_pass)
+          {
+            cluster_offset(update) = cluster_sizes(i);
+            map_cluster_to_offset_position(i) = update;
+          }
+          ++update;
+        }
+        else
+        {
+          if (final_pass)
+            map_cluster_to_offset_position(i) = IGNORED_CLUSTER;
+        }
+      },
+      num_clusters);
+  Kokkos::resize(Kokkos::WithoutInitializing, cluster_offset, num_clusters + 1);
+  KokkosExt::exclusive_scan(exec_space, cluster_offset, cluster_offset, 0);
+
+  auto cluster_starts = KokkosExt::clone(exec_space, cluster_offset);
+  KokkosExt::reallocWithoutInitializing(
+      exec_space, cluster_indices,
+      KokkosExt::lastElement(exec_space, cluster_offset));
+  Kokkos::parallel_for(
+      "ArborX::DBSCAN::compute_cluster_indices",
+      Kokkos::RangePolicy(exec_space, 0, n), KOKKOS_LAMBDA(int const i) {
+        // Ignore noise points
+        if (labels(i) < 0)
+          return;
+
+        auto offset_pos = map_cluster_to_offset_position(labels(i));
+        if (offset_pos != IGNORED_CLUSTER)
+        {
+          auto position =
+              Kokkos::atomic_fetch_add(&cluster_starts(offset_pos), 1);
+          cluster_indices(position) = i;
+        }
+      });
+
+  Kokkos::Profiling::popRegion();
+}
+
+template <typename ExecutionSpace, typename Primitives>
+bool run_dbscan(ExecutionSpace const &exec_space, Primitives const &primitives,
+                ArborXBenchmark::Parameters const &params)
+{
+  using MemorySpace = typename Primitives::memory_space;
+
+  if (params.verbose)
+  {
+    Kokkos::Profiling::Experimental::set_push_region_callback(
+        ArborX_Benchmark::push_region);
+    Kokkos::Profiling::Experimental::set_pop_region_callback(
+        ArborX_Benchmark::pop_region);
+  }
+
+  Kokkos::View<int *, MemorySpace> labels("Example::labels", 0);
+  using ArborX::DBSCAN::Implementation;
+  Implementation implementation = Implementation::FDBSCAN;
+  if (params.implementation == "fdbscan-densebox")
+    implementation = Implementation::FDBSCAN_DenseBox;
+
+  ArborX::DBSCAN::Parameters dbscan_params;
+  dbscan_params.setVerbosity(params.verbose).setImplementation(implementation);
+
+  Kokkos::Profiling::pushRegion("ArborX::DBSCAN::total");
+
+  labels = ArborX::dbscan<ExecutionSpace>(exec_space, primitives, params.eps,
+                                          params.core_min_size, dbscan_params);
+
+  Kokkos::Profiling::pushRegion("ArborX::DBSCAN::postprocess");
+  Kokkos::View<int *, MemorySpace> cluster_indices("Testing::cluster_indices",
+                                                   0);
+  Kokkos::View<int *, MemorySpace> cluster_offset("Testing::cluster_offset", 0);
+  sortAndFilterClusters(exec_space, labels, cluster_indices, cluster_offset,
+                        params.cluster_min_size);
+  Kokkos::Profiling::popRegion();
+
+  Kokkos::Profiling::popRegion();
+
+  if (params.verbose)
+  {
+    bool const is_special_case = (params.core_min_size == 2);
+
+    if (implementation == ArborX::DBSCAN::Implementation::FDBSCAN_DenseBox)
+      printf("-- dense cells      : %10.3f\n",
+             ArborX_Benchmark::get_time("ArborX::DBSCAN::dense_cells"));
+    printf("-- construction     : %10.3f\n",
+           ArborX_Benchmark::get_time("ArborX::DBSCAN::tree_construction"));
+    printf("-- query+cluster    : %10.3f\n",
+           ArborX_Benchmark::get_time("ArborX::DBSCAN::clusters"));
+    if (!is_special_case)
+    {
+      printf("---- neigh          : %10.3f\n",
+             ArborX_Benchmark::get_time("ArborX::DBSCAN::clusters::num_neigh"));
+      printf("---- query          : %10.3f\n",
+             ArborX_Benchmark::get_time("ArborX::DBSCAN::clusters::query"));
+    }
+    printf("-- postprocess      : %10.3f\n",
+           ArborX_Benchmark::get_time("ArborX::DBSCAN::postprocess"));
+    printf("total time          : %10.3f\n",
+           ArborX_Benchmark::get_time("ArborX::DBSCAN::total"));
+  }
+
+  int num_points = primitives.extent_int(0);
+  int num_clusters = cluster_offset.size() - 1;
+  int num_cluster_points = cluster_indices.size();
+  printf("\n#clusters       : %d\n", num_clusters);
+  printf("#cluster points : %d [%.2f%%]\n", num_cluster_points,
+         (100.f * num_cluster_points / num_points));
+  int num_noise_points = num_points - num_cluster_points;
+  printf("#noise   points : %d [%.2f%%]\n", num_noise_points,
+         (100.f * num_noise_points / num_points));
+
+  bool success = true;
+  if (params.verify)
+  {
+    success = ArborX::Details::verifyDBSCAN(exec_space, primitives, params.eps,
+                                            params.core_min_size, labels);
+    printf("Verification %s\n", (success ? "passed" : "failed"));
+  }
+
+  if (success && !params.filename_labels.empty())
+    writeLabelsData(params.filename_labels, labels);
+
+  return success;
 }
 
 template <typename T>
@@ -69,30 +246,29 @@ int main(int argc, char *argv[])
 {
   Kokkos::ScopeGuard guard(argc, argv);
 
+  using ExecutionSpace = Kokkos::DefaultExecutionSpace;
+  using MemorySpace = ExecutionSpace::memory_space;
+
   std::cout << "ArborX version    : " << ArborX::version() << std::endl;
   std::cout << "ArborX hash       : " << ArborX::gitCommitHash() << std::endl;
   std::cout << "Kokkos version    : " << ArborX::Details::KokkosExt::version()
             << std::endl;
 
   namespace bpo = boost::program_options;
+  using namespace ArborXBenchmark;
 
-  ArborXBenchmark::Parameters params;
-  int dim;
+  Parameters params;
 
-  std::vector<std::string> allowed_algorithms = {"dbscan", "hdbscan", "mst"};
-  std::vector<std::string> allowed_dendrograms = {"boruvka", "union-find"};
   std::vector<std::string> allowed_impls = {"fdbscan", "fdbscan-densebox"};
 
   bpo::options_description desc("Allowed options");
   // clang-format off
   desc.add_options()
       ( "help", "help message" )
-      ( "algorithm", bpo::value<std::string>(&params.algorithm)->default_value("dbscan"), ("algorithm " + vec2string(allowed_algorithms, " | ")).c_str() )
       ( "binary", bpo::bool_switch(&params.binary), "binary file indicator")
       ( "cluster-min-size", bpo::value<int>(&params.cluster_min_size)->default_value(1), "minimum cluster size")
       ( "core-min-size", bpo::value<int>(&params.core_min_size)->default_value(2), "DBSCAN min_pts")
-      ( "dendrogram", bpo::value<std::string>(&params.dendrogram)->default_value("boruvka"), ("dendrogram " + vec2string(allowed_dendrograms, " | ")).c_str() )
-      ( "dimension", bpo::value<int>(&dim)->default_value(3), "dimension of points to generate" )
+      ( "dimension", bpo::value<int>(&params.dim)->default_value(-1), "dimension of points to generate" )
       ( "eps", bpo::value<float>(&params.eps), "DBSCAN eps" )
       ( "filename", bpo::value<std::string>(&params.filename), "filename containing data" )
       ( "impl", bpo::value<std::string>(&params.implementation)->default_value("fdbscan"), ("implementation " + vec2string(allowed_impls, " | ")).c_str() )
@@ -131,81 +307,41 @@ int main(int argc, char *argv[])
               << "\n";
     return 2;
   }
-  if (!found(allowed_algorithms, params.algorithm))
-  {
-    std::cerr << "Algorithm must be one of " << vec2string(allowed_algorithms)
-              << "\n";
-    return 3;
-  }
-  if (!found(allowed_dendrograms, params.dendrogram))
-  {
-    std::cerr << "Dendrogram must be one of " << vec2string(allowed_dendrograms)
-              << "\n";
-    return 4;
-  }
-
-  std::stringstream ss;
-  ss << params.implementation;
 
   // Print out the runtime parameters
-  printf("algorithm         : %s\n", params.algorithm.c_str());
-  if (params.algorithm == "dbscan")
-  {
-    printf("eps               : %f\n", params.eps);
-    printf("cluster min size  : %d\n", params.cluster_min_size);
-    printf("implementation    : %s\n", ss.str().c_str());
-    printf("verify            : %s\n", (params.verify ? "true" : "false"));
-  }
-  if (params.algorithm == "hdbscan")
-  {
-    printf("dendrogram        : %s\n", params.dendrogram.c_str());
-  }
+  std::stringstream ss;
+  ss << params.implementation;
+  printf("eps               : %f\n", params.eps);
   printf("minpts            : %d\n", params.core_min_size);
-  if (!params.filename.empty())
-  {
-    // Data is read in
-    printf("filename          : %s [%s, max_pts = %d]\n",
-           params.filename.c_str(), (params.binary ? "binary" : "text"),
-           params.max_num_points);
-    printf("samples           : %d\n", params.num_samples);
-  }
-  else
-  {
-    // Data is generated
-    printf("generator         : n = %d, dim = %d, density = %s\n", params.n,
-           dim, (params.variable_density ? "variable" : "constant"));
-  }
+  printf("cluster min size  : %d\n", params.cluster_min_size);
   if (!params.filename_labels.empty())
     printf("filename [labels] : %s [binary]\n", params.filename_labels.c_str());
+  printf("implementation    : %s\n", ss.str().c_str());
+  printf("verify            : %s\n", (params.verify ? "true" : "false"));
   printf("verbose           : %s\n", (params.verbose ? "true" : "false"));
 
-  if (!params.filename.empty())
-    dim = getDataDimension(params.filename, params.binary);
+  ExecutionSpace exec_space;
 
-  using ArborXBenchmark::run;
-
-  bool success;
+  int dim = (params.filename.empty()
+                 ? params.dim
+                 : getDataDimension(params.filename, params.binary));
+#define SWITCH_DIM(DIM)                                                        \
+  case DIM:                                                                    \
+    success =                                                                  \
+        run_dbscan(exec_space, loadData<DIM, MemorySpace>(params), params);    \
+    break;
+  bool success = true;
   switch (dim)
   {
-  case 2:
-    success = run<2>(params);
-    break;
-  case 3:
-    success = run<3>(params);
-    break;
-  case 4:
-    success = run<4>(params);
-    break;
-  case 5:
-    success = run<5>(params);
-    break;
-  case 6:
-    success = run<6>(params);
-    break;
+    SWITCH_DIM(2)
+    SWITCH_DIM(3)
+    SWITCH_DIM(4)
+    SWITCH_DIM(5)
+    SWITCH_DIM(6)
   default:
     std::cerr << "Error: dimension " << dim << " not allowed\n" << std::endl;
-    success = false;
   }
+#undef SWITCH_DIM
 
   return success ? EXIT_SUCCESS : EXIT_FAILURE;
 }
